Journals and References/YOLO/scripts/test_requirements.txt
# Requirements for test_services.py
httpx==0.26.0
redis==5.0.1
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
